requirements.txt
torch>=1.12.0
triton>=2.0.0
cupy-cuda11x>=11.0.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
pydantic>=1.8.0
transformers>=4.20.0
scikit-learn>=1.0.0
requests>=2.25.0
matplotlib>=3.5.0
argparse
